+++ Example/SwiftySoundExample/ViewController.swift
@@ -12,6 +12,7 @@ import SwiftySound
 class ViewController: UIViewController, UITextFieldDelegate {
 
     @IBOutlet weak var switchSoundEnabled: UISwitch!
+    @IBOutlet weak var switchBackgroundMusic: UISwitch!
     @IBOutlet weak var textFieldNumberOfLoops: UITextField!
     @IBOutlet weak var buttonDog: UIButton!
     @IBOutlet weak var buttonCat: UIButton!
@@ -20,6 +21,7 @@ class ViewController: UIViewController, UITextFieldDelegate {
     @IBOutlet weak var sliderVolume: UISlider!
     @IBOutlet weak var buttonDogWithVolume: UIButton!
 
+    private var backgroundSound: Sound?
     private var dogSound: Sound?
 
     override func viewDidLoad() {
@@ -27,6 +29,11 @@ class ViewController: UIViewController, UITextFieldDelegate {
         switchSoundEnabled.isOn = Sound.enabled
         if let dogUrl = Bundle.main.url(forResource: "dog", withExtension: "wav") {
             dogSound = Sound(url: dogUrl)
+        }
+        if let pianoUrl = Bundle.main.url(forResource: "piano", withExtension: "wav") {
+            backgroundSound = Sound(url: pianoUrl)
+            backgroundSound?.volume = 0.8
+            backgroundSound?.prepare()
         }
     }
 
@@ -54,15 +61,40 @@ class ViewController: UIViewController, UITextFieldDelegate {
 
     @IBAction func buttonStopClicked(_ sender: Any) {
         Sound.stopAll()
+        switchBackgroundMusic.isOn = false
     }
 
     @IBAction func switchSoundEnabledValueChanged(_ sender: UISwitch) {
         Sound.enabled = sender.isOn
+        if switchBackgroundMusic.isOn {
+            enableBackgroundMusic()
+        }
+    }
+
+    @IBAction func switchBackgroundMusicValueChanged(_ sender: UISwitch) {
+        if sender.isOn {
+            enableBackgroundMusic()
+        }
+        else {
+            disableBackgroundMusic()
+        }
     }
 
     @IBAction func sliderVolumeValueChanged(_ sender: UISlider) {
         dogSound?.volume = sender.value
         labelVolume.text = "volume: \(String(format: "%0.0f", (sender.value * 100)))%"
+    }
+
+    // MARK: - Background music
+    private func enableBackgroundMusic() {
+        guard let backgroundSound = backgroundSound else { return }
+        if !backgroundSound.resume() { // trying to resume
+            backgroundSound.play(numberOfLoops: -1) // couldn't resume. most likely because it hasn't been played yet. playing the background sound from start. negative number of loops will cause the sound to loop infinitely
+        }
+    }
+
+    private func disableBackgroundMusic() {
+        backgroundSound?.pause()
     }
 
     // MARK: - UITextFieldDelegate